requirements.txt
numpy~=1.26.1
pandas~=2.1.1
Flask~=3.0.0
scikit-learn
requests~=2.31.0
Pillow~=10.1.0
gunicorn == 20.0.4
torch~=2.1.0
torchvision~=0.16.0
config~=0.5.1
MarkupSafe~=2.1.3